src/sync/requirements.txt
elasticsearch==8.12.0
mysql-connector-python